requirements.txt
numpy>=1.26,<2
Pillow>=10
tqdm>=4.66
torch==2.7.1
torchvision==0.22.1
matplotlib>=3.9
